+++ CHANGELOG.md
@@ -27,7 +27,7 @@
 ### Fix
 
 - **flakehell**: fix and clean config
-- **models**: change clean_name() to use the titlecase librairy
+- **models**: change clean_name() to use the titlecase library
 - update typing information for Python 3.9 and mypy 0.800
 
 ### Feat
@@ -44,7 +44,7 @@
 ### Fix
 
 - **flakehell**: fix and clean config
-- **models**: change clean_name() to use the titlecase librairy
+- **models**: change clean_name() to use the titlecase library
 - update typing information for Python 3.9 and mypy 0.800
 
 ## v0.5.0 (2021-02-06)